requirements.txt
Flask==3.0.0
flask-cors==4.0.0
google-generativeai==0.3.2
requests==2.31.0
gunicorn==21.2.0
